requirements-dev.txt
-r requirements.txt
astroid==2.15.8
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
filelock==3.18.0
flake8==6.1.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-django==1.4
flake8-quotes==3.4.0
identify==2.6.10
isort==6.0.1
lazy-object-proxy==1.11.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyYAML==6.0.2
setuptools==80.8.0
virtualenv==20.31.2
wrapt==1.17.2